Fix most of the warnings

+++ tests/test_insert_chunk.c
@@ -151,9 +151,9 @@ static char* test_insert_chunk(void) {
 
 static char *all_tests(void) {
 
-  for (int i = 0; i < sizeof(tstorage) / sizeof(test_storage); ++i) {
-    for (int j = 0; j < sizeof(tndata) / sizeof(test_ndata); ++j) {
-      for (int k = 0; k < sizeof(tcopy) / sizeof(bool); ++k) {
+  for (int i = 0; i < (int) (sizeof(tstorage) / sizeof(test_storage)); ++i) {
+    for (int j = 0; j < (int) (sizeof(tndata) / sizeof(test_ndata)); ++j) {
+      for (int k = 0; k < (int) (sizeof(tcopy) / sizeof(bool)); ++k) {
 
         tdata.contiguous = tstorage[i].contiguous;
         tdata.urlpath = tstorage[i].urlpath;
